pyproject.toml
[tool.poetry]
name = "yousim"
version = "0.1.0"
description = "YouSim: Using Claude 3 Opus to Simulate identities located in the latent space"
authors = ["vintro <vince@plasticlabs.ai>", "vineeth <vineeth@plasticlabs.ai>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
mirascope = {extras = ["anthropic"], version = "^0.16.0"}
honcho-ai = "^0.0.12"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
